% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score_calculator.R
\name{zscore}
\alias{zscore}
\title{Calculate Z-Scores from Count Tables}
\usage{
zscore(scored_table, control_table)
}
\arguments{
\item{scored_table}{Data frame of samples to be scored (genes as rows, samples as columns).}

\item{control_table}{Data frame of control samples (genes as rows, samples as columns).}
}
\value{
A data frame with the sum of Z-scores per sample and the sample IDs.
}
\description{
This function computes a Z-score sum for each sample in the given "scored" count table, based on
the means and SDs of the genes in the control table.
}
